requirements/base.txt
Django==2.0.6
django-model-utils==3.1.2
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2018.4
sorl-thumbnail==12.4.1
Unipath==1.1
